// src/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const PlayStatus = {
  ACTIVE: 'ACTIVE',
  CLOSED: 'CLOSED'
} as const

export type PlayStatus = (typeof PlayStatus)[keyof typeof PlayStatus]


export const RecordStatus = {
  LIKE: 'LIKE',
  DISLIKE: 'DISLIKE',
  EAT: 'EAT'
} as const

export type RecordStatus = (typeof RecordStatus)[keyof typeof RecordStatus]